Rename DataFile to Volume and use named struct initialization in tests
Rename DataFile to Volume to be consistent with the rest of the code.
The initial datafile name is confusing because the object server already
has "datafiles".

Named struct initialization now needed as the generated grpc code for
golang now adds XXX_ fields to the structs and that broke the tests.